To help prospective users understand the limitations
I was looking for something more powerful than what Github Copilot can currently do, but Copilot can turn a single \sum into a list comprehension and a sum in Python. Others may also be looking for different functionality, so it might be good to expand the description a bit, such as I propose here.